Added checks for only 1 query document result

+++ src/main/java/hu/kristof/nagy/hikebookserver/service/groups/GroupsGeneralConnectService.java
@@ -4,6 +4,7 @@ import com.google.cloud.firestore.Firestore;
 import hu.kristof.nagy.hikebookserver.data.DbPathConstants;
 import hu.kristof.nagy.hikebookserver.model.Group;
 import hu.kristof.nagy.hikebookserver.service.FutureUtil;
+import hu.kristof.nagy.hikebookserver.service.route.QueryException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -44,12 +45,21 @@ public class GroupsGeneralConnectService {
                 .whereEqualTo(DbPathConstants.GROUP_MEMBER_NAME, userName)
                 .get();
         return FutureUtil.handleFutureGet(() -> {
-            String id = queryFuture.get().getDocuments().get(0).getId();
-            db.collection(DbPathConstants.COLLECTION_GROUP)
-                    .document(id)
-                    .delete()
-                    .get();
-            return true;
+            var queryDocs = queryFuture.get().getDocuments();
+            if (queryDocs.size() > 1) {
+                throw new QueryException("Got more than 1 query document snapshot, but was only expecting 1. " +
+                        "Group name: " + groupName + ", member name: " + userName);
+            } else if (queryDocs.size() == 0) {
+                throw new QueryException("Got no query document snapshot, but was only expecting 1. " +
+                        "Group name: " + groupName + ", member name: " + userName);
+            } else {
+                String id = queryDocs.get(0).getId();
+                db.collection(DbPathConstants.COLLECTION_GROUP)
+                        .document(id)
+                        .delete()
+                        .get();
+                return true;
+            }
         });
     }
 }
